+++ src/main/java/com/oop/CalculatorHistoryStub.java
@@ -1,0 +1,13 @@
+package com.oop;
+
+public class CalculatorHistoryStub implements CalculatorHistory {
+    @Override
+    public void addToLog(String command) {
+        System.out.println("[DEBUG] Command logged: '" + command + "'");
+    }
+
+    @Override
+    public String[] getLog() {
+        return new String[] { "[DEBUG] test entry 1", "[DEBUG] test entry 2" };
+    }
+}
