--- src/main/java/com/oop/CalculatorLoggerFactory.java
@@ -1,7 +1,0 @@
-package com.oop;
-
-public class CalculatorLoggerFactory {
-    public static CalculatorLogger create() {
-        return null;
-    }
-}
